pyproject.toml
[project]
name = "arcanalyse"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.111",
  "uvicorn[standard]>=0.30",
  "sqlmodel>=0.0.21",
  "sqlalchemy[asyncio]>=2.0.30",
  "psycopg[binary]>=3.1.19",
  "alembic>=1.13.2",
  "pydantic>=2.7",
  "pydantic-settings>=2.2",
  "python-dotenv>=1.0.1",
  "email-validator>=2.1",
  "asyncpg>=0.29",
  "argon2-cffi>=25.1.0",
]


[tool.uv]
dev-dependencies = [
  "black>=24.8",
  "ruff>=0.6.9",
  "mypy>=1.11",
  "pytest>=8.3",
  "pytest-asyncio>=0.23",
  "pytest-sugar>=1.0",
  "httpx>=0.27",
  "asgi-lifespan>=2.1",
  "pre-commit>=3.7",
]
# Zentraler Ort für Tooling-Konfiguration (keine Python-Package-Definition nötig)

[tool.black]
line-length = 100
target-version = ["py312"]
preview = true

[tool.ruff]
line-length = 100
target-version = "py312"
src = ["backend"]
extend-exclude = [".venv", "backend/.venv"]

[tool.ruff.lint]
# Minimal, aber sinnvoll: PEP8/Flake (E/F/W), Imports (I), Bugbear (B), pyupgrade (UP)
select = ["E", "F", "W", "I", "B", "UP"]
# Keine globalen Ignores – wir halten es streng und sauber

[tool.ruff.lint.isort]
known-first-party = ["backend"]

[tool.mypy]
python_version = "3.12"
mypy_path = ["backend"]
exclude = '(^|/)backend/tests/'
ignore_missing_imports = true
follow_imports = "silent"

disallow_untyped_defs = false
check_untyped_defs = false
no_implicit_optional = false
warn_return_any = false

disable_error_code = [
  "attr-defined",
  "redundant-cast",
  "unused-ignore",
  "misc"
]

warn_unused_ignores = false
warn_redundant_casts = false
pretty = true
show_error_codes = true

[tool.pytest.ini_options]
addopts = "-vv -rA --color=yes"
asyncio_mode = "strict"
testpaths = ["tests"]
pythonpath = ["backend"]
